New Enemies + change lot of int to float to properly work

+++ src/main/GameObject.java
@@ -4,11 +4,11 @@ import java.awt.*;
 
 public abstract class GameObject {
 
-    protected int x,y;
+    protected float x,y;
     protected ID id;
-    protected int velX, velY; //velocity
+    protected float velX, velY; //velocity
 
-    public GameObject(int x, int y, ID id){
+    public GameObject(float x, float y, ID id){
         this.x = x;
         this.y = y;
         this.id = id;
@@ -18,16 +18,16 @@ public abstract class GameObject {
     public abstract void render(Graphics g);
     public abstract Rectangle getBounds();
 
-    public void setVelX(int velX) {
+    public void setVelX(float velX) {
         this.velX = velX;
     }
-    public void setVelY(int velY) {
+    public void setVelY(float velY) {
         this.velY = velY;
     }
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
     public void setId(ID id) {
@@ -36,18 +36,18 @@ public abstract class GameObject {
     public ID getId(){
         return id;
     }
-    public int getX() {
+    public float getX() {
         return x;
     }
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public int getVelX() {
+    public float getVelX() {
         return velX;
     }
 
-    public int getVelY() {
+    public float getVelY() {
         return velY;
     }
 }
